% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimmi.R
\name{as.mids.vimmi}
\alias{as.mids.vimmi}
\title{Convert a vimmi object to a mice mids object}
\usage{
as.mids.vimmi(x, ...)
}
\arguments{
\item{x}{A \code{vimmi} object}

\item{...}{Currently unused}
}
\value{
A \code{mids} object (from the mice package)
}
\description{
Converts a \code{vimmi} object to long format and uses
\code{mice::as.mids()} to create a proper \code{mids} object.
This enables use of \code{mice::pool()}, \code{mice::with.mids()},
and other mice infrastructure.
}
