% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIM-package.R
\docType{data}
\name{collisions}
\alias{collisions}
\title{Subset of the collision data}
\source{
\url{https://data.cityofnewyork.us/Public-Safety/NYPD-Motor-Vehicle-Collisions/h9gi-nx95}
}
\description{
Subset of the collision data from December 20. to December 31. 2018 from NYCD.
}
\details{
Each record represents a collision in NYC by city, borough, precinct and cross street.
}
\examples{

data(collisions)
aggr(collisions)

}
\keyword{datasets}
