% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimids.R
\name{complete}
\alias{complete}
\title{Extract completed datasets}
\usage{
complete(data, ...)
}
\arguments{
\item{data}{A multiply imputed object}

\item{...}{Additional arguments}
}
\description{
Generic function to extract completed datasets from a multiply imputed object.
}
