% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeCellGLoc.R
\name{.gloc_bisquare}
\alias{.gloc_bisquare}
\title{Tukey bisquare weights for standardised conditional residuals}
\usage{
.gloc_bisquare(Z, c = 4.685)
}
\arguments{
\item{Z}{matrix of standardised conditional residuals.}

\item{c}{tuning constant; \code{Inf} disables downweighting.}
}
\value{
a matrix of weights in [0, 1] with the dimensions of \code{Z}.
}
\description{
Tukey bisquare weights for standardised conditional residuals
}
\keyword{internal}
