% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellgloc_categorical.R
\name{.gloc_cat_identity}
\alias{.gloc_cat_identity}
\title{The E-step result for data without missing categorical cells}
\usage{
.gloc_cat_identity(catp, U)
}
\arguments{
\item{catp}{result of \code{.gloc_cat_prepare}.}

\item{U}{the design.}
}
\description{
The E-step result for data without missing categorical cells
}
\keyword{internal}
