% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellgloc_categorical.R
\name{.gloc_chol_psd}
\alias{.gloc_chol_psd}
\title{A square root of a positive semi-definite matrix}
\usage{
.gloc_chol_psd(C)
}
\arguments{
\item{C}{symmetric positive semi-definite matrix.}
}
\value{
\code{R} with \code{t(R) \%*\% R} equal to \code{C}.
}
\description{
A square root of a positive semi-definite matrix
}
\keyword{internal}
