% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIM-package.R
\docType{data}
\name{food}
\alias{food}
\title{Food consumption}
\format{A data frame with 16 observations on the following 21 variables.}
\source{
\url{https://openmv.net/info/food-consumption}
}
\description{
The relative consumption of certain food items in European and Scandinavian countries.
}
\details{
The numbers represent the percentage of the population consuming that food type.
}
\examples{

data(food)
str(food)
aggr(food)

}
\keyword{datasets}
