% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxCat.R
\name{maxCat}
\alias{maxCat}
\title{Aggregation function for a factor variable}
\usage{
maxCat(x, weights = NULL)
}
\arguments{
\item{x}{factor vector}

\item{weights}{numeric vector providing weights for the observations in x}
}
\description{
The function maxCat chooses the level
with the most occurrences and random if the maximum is not unique.
}
