% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlpImp.R
\name{mlpImp}
\alias{mlpImp}
\title{Deep Learning Imputation for Categorical Variables}
\usage{
mlpImp(data, variable = colnames(data), modelVariables = colnames(data),
  imp_var = TRUE, imp_suffix = "imp")
}
\arguments{
\item{data}{data.frame, data.table, survey object or matrix}

\item{variable}{variables to be imputed}

\item{modelVariables}{variables used for modelling}

\item{imp_var}{TRUE/FALSE if a TRUE/FALSE variables for each imputed
variable should be created show the imputation status}

\item{imp_suffix}{suffix for the TRUE/FALSE variables showing the imputation
status}
}
\value{
the imputed data set.
}
\description{
mxnet is used
}
\details{
only works for character and factor variables
}
\examples{

data(sleep,package="VIM")
d2 <- sleep#kNN(sleep,imp_var = FALSE)
d2$Pred <- as.character(d2$Pred)
d2$Exp <- as.factor(d2$Exp)
d2$Danger <- as.factor(d2$Danger)
d2[1:10,c("Pred","Danger")] <- NA
di <- mlpImp(d2)


}
\author{
Alexander Kowarik
}
