% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIM-package.R
\docType{data}
\name{toydataMiss}
\alias{toydataMiss}
\title{Simulated toy data set for examples}
\format{data frame with 100 observations and 12 variables. The first two 
variables represent the fully observed data.}
\description{
A 2-dimensional data set with additional information.
}
\examples{

data(toydataMiss)

}
\keyword{datasets}
