% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformerImpute.R
\name{transformerImpute}
\alias{transformerImpute}
\title{Transformer-based Imputation}
\usage{
transformerImpute(
  data,
  target,
  cat_vars = NULL,
  include_cols = NULL,
  imp_col = FALSE,
  num_unique_cat = 35,
  verbose = 1,
  dropout1 = 0.2,
  dropout2 = 0.3,
  num_heads = 8,
  num_transformer_blocks = 1,
  feed_forward_dim = 512,
  embedding_dim = 128,
  dense_dim = 128,
  epochs = 10,
  patience = 5,
  batch_size = 64,
  learning_rate = 1e-04,
  patience_min_delta = 0.01,
  regularization_factor = 0.01,
  monitor = "val_loss",
  decimal_points = 1,
  decimal_points_target = 1,
  round_long_decs = TRUE
)
}
\arguments{
\item{data}{Dataset to impute.}

\item{target}{Target variable to impute.}

\item{cat_vars}{Categorical variables.}

\item{include_cols}{Columns to include in imputation.}

\item{imp_col}{Whether to add an imputation indicator column.}

\item{num_unique_cat}{Number of unique values to consider as categorical.}

\item{verbose}{Verbosity level.}

\item{dropout1}{Dropout rate in transformer heads.}

\item{dropout2}{Dropout rate in feed forward network.}

\item{num_heads}{Number of attention heads per transformer block.}

\item{num_transformer_blocks}{Number of transformer blocks.}

\item{feed_forward_dim}{Dimension of feed forward network.}

\item{embedding_dim}{Dimension of embedding layer.}

\item{dense_dim}{Dimension of dense layer with relu activation.}

\item{epochs}{Number of epochs to train.}

\item{patience}{Number of epochs with no improvement to stop training.}

\item{batch_size}{Batch size for training.}

\item{learning_rate}{Learning rate for optimizer.}

\item{patience_min_delta}{Minimum change to qualify as improvement.}

\item{regularization_factor}{Regularization factor for kernel regularization.}

\item{monitor}{Evaluation measure to monitor.}

\item{decimal_points}{Number of decimal points for numeric input variables}

\item{decimal_points_target}{Number of decimal points for numeric target (ignored if target is categorical)}

\item{round_long_decs}{boolean: rounds numeric values with over 5 decimal points to only three decimal points (only if decimal_points is NULL)}
}
\value{
Imputed dataset.
}
\description{
Impute missing values using a transformer model.
}
