% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimpute_registry.R
\name{unregister_vimpute_method}
\alias{unregister_vimpute_method}
\title{Remove a user-registered \code{vimpute()} method}
\usage{
unregister_vimpute_method(name)
}
\arguments{
\item{name}{Single character string: the method name to remove. Built-in
methods cannot be removed.}
}
\value{
Invisibly \code{TRUE} on success.
}
\description{
Remove a user-registered \code{vimpute()} method
}
\examples{
register_vimpute_method("cart",
  learner = list(regr = "regr.rpart", classif = "classif.rpart"))
unregister_vimpute_method("cart")
}
\seealso{
\code{\link[=register_vimpute_method]{register_vimpute_method()}}, \code{\link[=vimpute_methods]{vimpute_methods()}}

Other vimpute method registry: 
\code{\link{register_vimpute_method}()},
\code{\link{vimpute_methods}()}
}
\concept{vimpute method registry}
