% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableImportance.R
\name{variableImportance}
\alias{variableImportance}
\title{Variable Importance Function}
\usage{
variableImportance(X, Y, numTrees, sampleSize, m_feature, min_leaf, Command)
}
\arguments{
\item{X}{Training Features}

\item{Y}{Output Features}

\item{numTrees}{Number of Trees to generate}

\item{sampleSize}{Size of random subset for each tree generation}

\item{m_feature}{Number of randomly selected features considered for a split in each regression tree node, which
must be positive integer and less than N (number of input features)}

\item{min_leaf}{Minimum number of samples in the leaf node. If a node has less than or equal
to min_leaf samples, then there will be no splitting in that node and this node
will be considered as a leaf node. Valid input is positive integer, which is less
than or equal to M (number of training samples)}

\item{Command}{1 for univariate Regression Tree (corresponding to RF) and 2 for Multivariate
Regression Tree (corresponding to MRF)}
}
\value{
Vector of size N x 1
}
\description{
Variable Importance Function
}
\examples{
X = matrix(runif(50*100), 50, 100)
Y = matrix(runif(50*5), 50, 5)
m_feature = 5
min_leaf = 5
n_trees = 100
sampleSize = 20
Command = 2
variableImportance(X, Y, n_trees, sampleSize, m_feature, min_leaf, Command)
}
