% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimpute_registry.R
\name{vimpute_methods}
\alias{vimpute_methods}
\title{List the imputation methods registered for \code{vimpute()}}
\usage{
vimpute_methods()
}
\value{
Character vector of registered method names (built-ins plus any
methods added via \code{\link[=register_vimpute_method]{register_vimpute_method()}}).
}
\description{
List the imputation methods registered for \code{vimpute()}
}
\examples{
vimpute_methods()
}
\seealso{
\code{\link[=register_vimpute_method]{register_vimpute_method()}}, \code{\link[=unregister_vimpute_method]{unregister_vimpute_method()}}

Other vimpute method registry: 
\code{\link{register_vimpute_method}()},
\code{\link{unregister_vimpute_method}()}
}
\concept{vimpute method registry}
