% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellImpForest.R
\name{cellImpForest}
\alias{cellImpForest}
\title{Cellwise-robust imputation of mixed data with tree ensembles}
\usage{
cellImpForest(data, ...)
}
\arguments{
\item{data}{a data.frame with numeric and categorical columns, NA allowed}

\item{...}{further arguments (not yet used)}
}
\value{
an object of class \code{cellImpForest}
}
\description{
Documentation is completed in Task 7.
}
