% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeCellGLoc.R
\name{.gloc_scatter_soft}
\alias{.gloc_scatter_soft}
\title{Cellwise weighted maximum likelihood scatter of the residuals}
\usage{
.gloc_scatter_soft(R, W, M)
}
\arguments{
\item{R}{\eqn{n x p} matrix of residuals from the mean structure.}

\item{W}{\eqn{n x p} matrix of cell weights in [0, 1].}

\item{M}{\eqn{n x p} logical mask of missing cells.}
}
\value{
a \eqn{p x p} scatter matrix.
}
\description{
Cellwise weighted maximum likelihood scatter of the residuals
}
\keyword{internal}
