% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impPCA.R
\name{impPCA}
\alias{impPCA}
\title{Iterative EM PCA imputation}
\usage{
impPCA(
  x,
  method = "classical",
  m = 1,
  eps = 0.5,
  k = ncol(x) - 1,
  maxit = 10,
  boot = FALSE
)
}
\arguments{
\item{x}{data.frame or matrix}

\item{method}{\code{"classical"} or \code{"mcd"} (robust estimation)}

\item{m}{number of multiple imputations (only if parameter \code{boot} equals \code{TRUE})}

\item{eps}{threshold for convergence}

\item{k}{number of principal components for reconstruction of \code{x}}

\item{maxit}{maximum number of iterations}

\item{boot}{residual bootstrap (if \code{TRUE})}
}
\value{
the imputed data set. If \code{boot = FALSE} this is a data.frame.
If \code{boot = TRUE} this is a list where each list element contains a data.frame.
}
\description{
Greedy algorithm for EM-PCA including robust methods
}
\examples{

data(Animals, package = "MASS")
Animals$brain[19] <- Animals$brain[19] + 0.01
Animals <- log(Animals)
colnames(Animals) <- c("log(body)", "log(brain)")
Animals_na <- Animals
probs <- abs(Animals$`log(body)`^2)
probs <- rep(0.5, nrow(Animals))
probs[c(6,16,26)] <- 0
set.seed(1234)
Animals_na[sample(1:nrow(Animals), 10, prob = probs), "log(brain)"] <- NA
impPCA(Animals_na)
impPCA(Animals_na, boot = TRUE, mi = 10)


}
\references{
Serneels, Sven and Verdonck, Tim (2008).
Principal component analysis for data containing outliers and missing elements.
Computational Statistics and Data Analysis, Elsevier, vol. 52(3), pages 1712-1727
}
\seealso{
Other imputation methods: 
\code{\link{hotdeck}()},
\code{\link{irmi}()},
\code{\link{kNN}()},
\code{\link{matchImpute}()},
\code{\link{rangerImpute}()},
\code{\link{regressionImp}()}
}
\author{
Matthias Templ
}
\concept{imputation methods}
\keyword{manip}
