% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeRobustChain.R
\name{imputeRobustChain}
\alias{imputeRobustChain}
\title{FUNCTION_TITLE}
\usage{
imputeRobustChain(
  formulas = vector("list", ncol(data)),
  data,
  boot = TRUE,
  robustboot = TRUE,
  method = "lts",
  multinom.method = "multinom",
  takeAll = TRUE,
  eps = 0.5,
  maxit = 4,
  alpha = 0.5,
  uncert = "pmm",
  familiy = "Gaussian",
  value_back = "matrix",
  trace = FALSE
)
}
\arguments{
\item{formulas}{PARAM_DESCRIPTION, Default: vector("list", ncol(data))}

\item{data}{PARAM_DESCRIPTION}

\item{boot}{PARAM_DESCRIPTION, Default: TRUE}

\item{robustboot}{PARAM_DESCRIPTION, Default: TRUE}

\item{method}{PARAM_DESCRIPTION, Default: 'lts'}

\item{multinom.method}{PARAM_DESCRIPTION, Default: 'multinom'}

\item{takeAll}{PARAM_DESCRIPTION, Default: TRUE}

\item{eps}{PARAM_DESCRIPTION, Default: 0.5}

\item{maxit}{PARAM_DESCRIPTION, Default: 4}

\item{alpha}{PARAM_DESCRIPTION, Default: 0.5}

\item{uncert}{PARAM_DESCRIPTION, Default: 'pmm'}

\item{familiy}{PARAM_DESCRIPTION, Default: 'Gaussian'}

\item{value_back}{PARAM_DESCRIPTION, Default: 'matrix'}

\item{trace}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[VIM]{initialise}}
\code{\link[robCompositions]{outCoDa}}
}
