\name{kola.background}
\alias{kola.background}
\docType{data}
\title{Background map for the Kola project data}
\description{
    Coordinates of the Kola background map.
}
\usage{
data(kola.background)
}
\source{
Kola Project (1993-1998)
}
\references{
Reimann, C., Filzmoser, P., Garrett, R.G. and Dutter, R. (2008) 
\emph{Statistical Data Analysis Explained: Applied Environmental
Statistics with R}. Wiley, 2008.
}
\examples{
data(kola.background, package = "VIM")
bgmap(kola.background)
}
\keyword{datasets}
