% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIM-package.R
\docType{data}
\name{lse_synthetic_rules}
\alias{lse_synthetic_rules}
\title{Validation rules for the synthetic LSE data}
\format{
A named list with three \code{validate::validator} objects:
\code{accounting}, \code{extra}, and \code{edit}.
}
\description{
Accounting, additional edit, and combined validation rules for
\link{lse_synthetic}.
}
\examples{

data(lse_synthetic_rules)
names(lse_synthetic_rules)

}
\keyword{datasets}
