% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleCat.R
\name{sampleCat}
\alias{sampleCat}
\title{Random aggregation function for a factor variable}
\usage{
sampleCat(x, weights = NULL)
}
\arguments{
\item{x}{factor vector}

\item{weights}{numeric vector providing weights for the observations in x}
}
\description{
The function sampleCat samples with probabilites corresponding to the
occurrence of the level in the NNs.
}
\seealso{
Other imputation methods: 
\code{\link{hotdeck}()},
\code{\link{impPCA}()},
\code{\link{irmi}()},
\code{\link{kNN}()},
\code{\link{matchImpute}()},
\code{\link{medianSamp}()},
\code{\link{rangerImpute}()},
\code{\link{regressionImp}()}
}
\concept{imputation methods}
