% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIM-package.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{Simulated data set for testing purpose}
\format{The format is: List of 4 $ wna :'data.frame': 500 obs. of 8
variables: ..$ x1: num [1:500] 10.87 9.53 7.83 8.53 8.67 ...  ..$ x2: num
[1:500] 10.9 9.32 7.68 8.2 8.41 ...  ..$ c1: Factor w/ 4 levels
"a","b","c","d": 3 2 2 1 2 2 1 3 3 2 ...  ..$ c2: Factor w/ 4 levels
"a","b","c","d": 2 3 2 2 2 2 2 4 2 2 ...  ..$ b1: Factor w/ 2 levels
"0","1": 2 2 1 2 1 2 1 2 1 1 ...  ..$ b2: Factor w/ 2 levels "0","1": 2 2 1
1 1 1 1 2 2 2 ...  ..$ m1: num [1:500] 0 8.29 9.08 0 0 ...  ..$ m2: num
[1:500] 10.66 9.39 7.8 8.11 7.33 ...  $ wona :'data.frame': 500 obs. of 8
variables: ..$ x1: num [1:500] 10.87 9.53 7.83 8.53 8.67 ...  ..$ x2: num
[1:500] 10.9 9.32 7.68 8.2 8.41 ...  ..$ c1: Factor w/ 4 levels
"a","b","c","d": 3 2 2 1 2 2 1 3 3 2 ...  ..$ c2: Factor w/ 4 levels
"a","b","c","d": 2 3 2 2 2 2 2 4 2 2 ...  ..$ b1: Factor w/ 2 levels
"0","1": 2 2 1 2 1 2 1 2 1 1 ...  ..$ b2: Factor w/ 2 levels "0","1": 2 2 1
1 1 1 1 2 2 2 ...  ..$ m1: num [1:500] 0 8.29 9.08 0 0 ...  ..$ m2: num
[1:500] 10.66 9.39 7.8 8.11 7.33 ...  $ mixed : chr [1:2] "m1" "m2" $
outlierInd: NULL}
\description{
2 numeric, 2 binary, 2 nominal and 2 mixed (semi-continous) variables
}
\examples{

data(testdata)

}
\keyword{datasets}

