% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimpute.R
\name{vimpute}
\alias{vimpute}
\title{VIM Model Based Imputation framework
generally idea function that incorporates a framework for model based imputation.
Options should be
\itemize{
\item sequential modelling
\item using PMM or not
\item bootstrap the model error
\item drawing the predicted value from the "posterior" distribution
\item model options: (robust) regression, ranger, XGBoost, some kind of transformer model
\item complex formula (with \code{formula}) for each variable possible.
}}
\usage{
vimpute(
  data,
  variable = colnames(data),
  sequential = TRUE,
  modeluncertainty = c("none", "bootstrap", "robustBootrap-stratified",
    "robustBootstrap-xyz", "BayesianBootstrap"),
  imputationuncertainty = "PMM",
  xvar = colnames(data),
  method = c("lm", "regression", "ranger", "XGBoost", "GPT"),
  formula = NULL,
  imp_var = FALSE,
  imp_suffix = "imp",
  verbose = FALSE
)
}
\arguments{
\item{data}{}

\item{variable}{}

\item{sequential}{}

\item{modeluncertainty}{}

\item{imputationuncertainty}{}

\item{xvar}{}

\item{method}{}

\item{formula}{}

\item{imp_var}{}

\item{imp_suffix}{}

\item{verbose}{}
}
\description{
VIM Model Based Imputation framework
generally idea function that incorporates a framework for model based imputation.
Options should be
\itemize{
\item sequential modelling
\item using PMM or not
\item bootstrap the model error
\item drawing the predicted value from the "posterior" distribution
\item model options: (robust) regression, ranger, XGBoost, some kind of transformer model
\item complex formula (with \code{formula}) for each variable possible.
}
}
